@@ -5,22 +5,22 @@ author:
 homepage:
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.19.6 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_styled_toast: ^1.4.0+1
+  flutter_styled_toast: ^2.2.1
   dio: 4.0.0-beta6
-  pull_to_refresh: 1.6.4
+  pull_to_refresh: ^2.0.0
   cached_network_image: ^3.0.0-nullsafety
   image_picker: ^0.7.2
   shared_preferences: ^2.0.3
   flutter_svg: ^0.21.0-nullsafety.0
   url_launcher: 6.0.3
-  shimmer: 1.1.2
+  shimmer: 3.0.0
   photo_view: 0.14.0
-  loading_indicator: ^1.2.0
+  loading_indicator: ^3.1.1
   waterfall_flow: 3.0.1
 
 dev_dependencies:
