@@ -15,10 +15,10 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.0+1
+version: 3.0.1+157
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.19.6 <3.0.0"
 
 dependencies:
   flutter:
@@ -30,14 +30,14 @@ dependencies:
   paixs_utils:
     path: package/paixs_utils
   provider: ^5.0.0
-  package_info: 0.4.1
+  package_info_plus: 8.3.0
   open_file: ^3.0.3
-  progress_dialog: 1.2.0
+  sn_progress_dialog: ^1.2.0
   permission_handler: ^7.1.0  #权限
   path_provider: 2.0.1
   # 轮播图
-  flutter_swiper: ^1.1.4
-  flutter_datetime_picker: 1.5.0
+  card_swiper: 3.0.1
+  flutter_datetime_picker: 1.5.1
 
   # 加密包，如sha1
   crypto: ^3.0.1
