@@ -144,21 +144,21 @@ Widget buildTFView(
   bool isExp = false,
   TextInputType keyboardType = TextInputType.text,
   bool obscureText = false,
-  @required String hintText,
-  void Function(String) onChanged,
-  void Function() onTap,
-  void Function(String) onSubmitted,
-  TextStyle textStyle,
-  TextStyle hintStyle,
-  FocusNode focusNode,
-  TextInputAction textInputAction,
+  required String hintText,
+  void Function(String)? onChanged,
+  void Function()? onTap,
+  void Function(String)? onSubmitted,
+  TextStyle? textStyle,
+  TextStyle? hintStyle,
+  FocusNode? focusNode,
+  TextInputAction? textInputAction,
   double hintSize = 14,
-  Color hintColor,
+  Color? hintColor,
   double textSize = 14,
-  Color textColor,
-  EdgeInsetsGeometry padding,
-  TextEditingController con,
-  TextAlign textAlign,
+  Color? textColor,
+  EdgeInsetsGeometry? padding,
+  TextEditingController? con,
+  TextAlign? textAlign,
   bool isInt = false,
   bool isDouble = false,
   bool isAz = false,
@@ -233,229 +233,3 @@ Widget buildTFView(
     )
   ][isExp ? 0 : 1];
 }
-
-
-// Future buildShowModalBottomSheet(
-//     BuildContext context,
-//     String id,
-//     String userId,
-//     GlobalKey _globalKey,
-//     String title,
-//     String description,
-//     ShareType shareType) {
-//   return showModalBottomSheet(
-//     context: context,
-//     builder: (_) => Container(
-//       height: 160,
-//       width: double.infinity,
-//       color: Colors.white,
-//       child: Column(
-//         mainAxisAlignment: MainAxisAlignment.start,
-//         children: <Widget>[
-//           SizedBox(
-//             height: 10,
-//           ),
-//           Row(
-//             children: <Widget>[
-//               Expanded(child: Text("")),
-//               Text(
-//                 "分享",
-//                 style: TextStyle(color: Color(0xff333333), fontSize: 19),
-//               ),
-//               Expanded(
-//                   child: GestureDetector(
-//                 onTap: () {
-//                   pop(context);
-//                 },
-//                 child: Container(
-//                     alignment: Alignment.centerRight,
-//                     padding:
-//                         EdgeInsets.only(right: 15, left: 15, top: 8, bottom: 8),
-//                     margin: EdgeInsets.only(left: 80),
-//                     child: Text("关闭")),
-//               )),
-//             ],
-//           ),
-//           Expanded(
-//             child: Container(
-//               alignment: Alignment.center,
-//               child: Row(
-//                 crossAxisAlignment: CrossAxisAlignment.center,
-//                 mainAxisAlignment: MainAxisAlignment.center,
-//                 children: <Widget>[
-//                   Column(
-//                     crossAxisAlignment: CrossAxisAlignment.center,
-//                     mainAxisAlignment: MainAxisAlignment.center,
-//                     children: <Widget>[
-//                       GestureDetector(
-//                         onTap: () {
-//                           Map webPath = {
-//                             ShareType.BENDI: Config.WebBaseUrl +
-//                                 "web/information/getInformation?id=$id&appUid=$userId",
-//                             ShareType.GOOD: Config.WebBaseUrl +
-//                                 "web/information/getGoodsDetail?id=$id&appUid=$userId",
-//                             ShareType.STORE: Config.WebBaseUrl +
-//                                 "web/information/getShopDetail?id=$id&appUid=$userId",
-//                             ShareType.RED: Config.WebBaseUrl +
-//                                 "web/information/getAreaInformation?id=$id&appUid=$userId",
-//                             ShareType.FRIST: Config.WebBaseUrl +
-//                                 "web/index/index?appUid=$userId"
-//                           };
-//                           Map miniPath = {
-//                             ShareType.BENDI:
-//                                 "pages/information/index?appUid=$userId&id=$id&source=detail",
-//                             ShareType.GOOD:
-//                                 "pages/goodsDetail/index?appUid=$userId&id=$id",
-//                             ShareType.STORE:
-//                                 "pages/shopDetail/index?appUid=$userId&id=$id",
-//                             ShareType.RED:
-//                                 "pages/areaInformation/index?appUid=$userId&id=$id&source=detail",
-//                             ShareType.FRIST: "pages/index/index?appUid=$userId"
-//                           };
-
-//                           if (shareType == ShareType.FRIST) {
-//                             // print("----------666-----------");
-//                             var model = new WeChatShareMiniProgramModel(
-//                                 webPageUrl: webPath[shareType],
-//                                 miniProgramType: WXMiniProgramType.RELEASE,
-//                                 userName: "gh_6e1e0c36ad7c",
-//                                 hdImagePath: WeChatImage.asset(
-//                                     "assets/img/shareimg1.jpg"),
-//                                 withShareTicket: true,
-//                                 path: miniPath[shareType],
-//                                 title: title,
-//                                 description: description,
-//                                 thumbnail: WeChatImage.asset(
-//                                     "assets/img/shareimg1.jpg"));
-
-//                             shareToWeChat(model);
-                            
-//                           } else {
-//                             capturePng(_globalKey).then((value) async {
-//                               Uint8List list = Uint8List.fromList(
-//                                   Luban.comressImageJpg(value, 50));
-//                               WeChatImage img = WeChatImage.binary(list);
-
-//                               var model1 = WeChatShareWebPageModel(
-//                                   "https://m.cngky.com/tg.html",
-//                                   title: "工控云物联网平台",
-//                                   description: "平台满足放租工厂、企业基础的定向设备物联租赁需求，如生产监控、设备启动、设备定位（设备超距）、设备销售、设备故障上报、设备续费、工单处理、客户关系维护、商户入驻、设备信息查询",
-//                                   thumbnail: img,
-//                                   scene: WeChatScene.SESSION);
-
-//                               // var model = new WeChatShareMiniProgramModel(
-//                               //     webPageUrl: webPath[shareType],
-//                               //     miniProgramType: WXMiniProgramType.RELEASE,
-//                               //     userName: "gh_6e1e0c36ad7c",
-//                               //     hdImagePath: img,
-//                               //     withShareTicket: true,
-//                               //     path: miniPath[shareType],
-//                               //     title: title,
-//                               //     description: description,
-//                               //     thumbnail: img);
-//                               shareToWeChat(model1).then((value){
-//                                       print("shareToWeChat$value");
-//                               });
-//                             });
-//                           }
-//                         },
-//                         child: Image.asset(
-//                           ImgConfig.weixin1,
-//                           width: 45,
-//                           height: 45,
-//                         ),
-//                       ),
-//                       SizedBox(
-//                         height: 3,
-//                       ),
-//                       Text(
-//                         "微信好友",
-//                         style:
-//                             TextStyle(fontSize: 13, color: Color(0xff999999)),
-//                       ),
-//                       SizedBox(
-//                         height: 7,
-//                       ),
-//                     ],
-//                   ),
-//                   SizedBox(
-//                     width: 40,
-//                   ),
-//                   Column(
-//                     crossAxisAlignment: CrossAxisAlignment.center,
-//                     mainAxisAlignment: MainAxisAlignment.center,
-//                     children: <Widget>[
-//                       GestureDetector(
-//                         onTap: () {
-//                           // print("----------777777-----------");
-//                           capturePng(_globalKey).then((value) {
-//                              Uint8List list = Uint8List.fromList(
-//                                   Luban.comressImageJpg(value, 50));
-//                               WeChatImage img = WeChatImage.binary(list);
-//                             // ignore: unused_local_variable
-//                             var model1 = new WeChatShareWebPageModel(
-//                               shareType == ShareType.BENDI
-//                                   ? Config.WebBaseUrl +
-//                                       "web/information/getInformation?id=$id&appUid=$userId"
-//                                   //  ?"http://m.ershengsan.cn/web/login/login?source=1&id=5ed08ec49a0be74b4dedf5a1&appUid=5d2177999a0be718ab32b9ec"
-//                                   : shareType == ShareType.GOOD
-//                                       ? Config.WebBaseUrl +
-//                                           "web/information/getGoodsDetail?id=$id&appUid=$userId"
-//                                       : shareType == ShareType.STORE
-//                                           ? Config.WebBaseUrl +
-//                                               "web/information/getShopDetail?id=$id&appUid=$userId"
-//                                           : shareType == ShareType.RED
-//                                               ? Config.WebBaseUrl +
-//                                                   "web/information/getAreaInformation?id=$id&appUid=$userId"
-//                                               : shareType == ShareType.FRIST
-//                                                   ? Config.WebBaseUrl +
-//                                                       "web/index/index?appUid=$userId"
-//                                                   : Config.WebBaseUrl +
-//                                                       "web/index/index?appUid=$userId",
-//                               scene: WeChatScene.TIMELINE,
-//                               title: title,
-//                               description: description,
-//                               thumbnail: shareType == ShareType.FRIST
-//                                   ? WeChatImage.asset("assets/img/gengmore.png")
-//                                   : WeChatImage.binary(value),
-//                             );
-
-                            
-//                               var model2 = WeChatShareWebPageModel(
-//                                   "https://m.cngky.com/tg.html",
-//                                   title: "工控云物联网平台",
-//                                   description: "平台满足放租工厂、企业基础的定向设备物联租赁需求，如生产监控、设备启动、设备定位（设备超距）、设备销售、设备故障上报、设备续费、工单处理、客户关系维护、商户入驻、设备信息查询",
-//                                   thumbnail: img,
-//                                   scene: WeChatScene.TIMELINE);
-//                             shareToWeChat(model2).then((value){
-//                               print("shareToWeChat$value");
-//                             });
-//                           });
-//                         },
-//                         child: Image.asset(
-//                           ImgConfig.pengyouquan1,
-//                           width: 43,
-//                           height: 43,
-//                         ),
-//                       ),
-//                       SizedBox(
-//                         height: 3,
-//                       ),
-//                       Text(
-//                         "朋友圈",
-//                         style:
-//                             TextStyle(fontSize: 13, color: Color(0xff999999)),
-//                       ),
-//                       SizedBox(
-//                         height: 7,
-//                       ),
-//                     ],
-//                   ),
-//                 ],
-//               ),
-//             ),
-//           ),
-//         ],
-//       ),
-//     ),
-//   );
